+++ Lab11/Lab10/Lab10/Controllers/CategoriesController.cs
@@ -140,6 +140,15 @@ namespace Lab10.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var category = await _context.Categories.FindAsync(id);
+            var articlesCount = await _context.Articles.Where(a => a.CategoryId == id).CountAsync();
+            if(articlesCount > 0)
+            {
+                ModelState.AddModelError(
+                    string.Empty,
+                    $"This category has {articlesCount} related articles. You can only delete categories with no articles."
+                    );
+                return View(category);
+            }
             _context.Categories.Remove(category);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
